Massively reduced the data usage.
The intro video and text is now part of the loading screen, reducing 
waiting time.

Improved the compression on the video a lot.

Sounds are no longer fetched in advance, improving load time.

webP is now used when availible for icons, also helping with network 
usage a lot.

Improved the logo.